requirements.txt
Flask==2.2.2
gunicorn==20.1.0
SQLAlchemy==1.4.27
Flask-Migrate==3.1.0
Flask-SQLAlchemy==2.5.1
Werkzeug==2.2.2
openai==1.58.1
twilio==9.4.1
python-dotenv==0.19.1
streamlit==1.35.0
